Move batch chat completion tests to dedicated test file
